AA, MO, Progression S, and cleanup
Added ShadowCharger  with the trainability of none because of block by fences.
Added Pigs from MO to have forage for Odyssey.
Removed comments from Mend and Recycle as it's been there for too long.
Added removal for Progression storage patch for Reel. Reel removed the ReelNeolithicStorage research job and needs to be updated in main mod. Steam commenter showed gist log to MA but no fix in place.